refactor: collapse datatype module into hl7v2
