+++ crates/hl7v2-datatype/Cargo.toml
@@ -13,12 +13,10 @@ keywords = ["hl7", "healthcare", "datatype", "validation"]
 categories = ["parser-implementations", "encoding"]
 
 [dependencies]
-regex = "1.0"
-thiserror = { workspace = true }
-chrono = { version = "0.4", features = ["serde", "std", "clock"], default-features = false }
-hl7v2-datetime = { version = "1.2.0", path = "../hl7v2-datetime" }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = ["profile"] }
 
 [dev-dependencies]
+chrono = { workspace = true }
 proptest = "1.6"
 cucumber = "0.22.1"
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
