+++ crates/hl7v2/Cargo.toml
@@ -25,8 +25,9 @@ json = ["dep:hl7v2-json"]
 profile = [
     "dep:hl7v2-prof",
     "dep:hl7v2-validation",
-    "dep:hl7v2-datatype",
     "dep:hl7v2-datetime",
+    "dep:chrono",
+    "dep:regex",
 ]
 ack = ["dep:hl7v2-ack"]
 normalize = ["dep:hl7v2-normalize"]
@@ -55,13 +56,13 @@ thiserror = { workspace = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 chrono = { workspace = true, features = ["serde"], optional = true }
 sha2 = { workspace = true, optional = true }
+regex = { workspace = true, optional = true }
 
 hl7v2-json = { version = "1.2.0", path = "../hl7v2-json", optional = true }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize", optional = true }
 hl7v2-ack = { version = "1.2.0", path = "../hl7v2-ack", optional = true }
 hl7v2-prof = { version = "1.2.0", path = "../hl7v2-prof", optional = true }
 hl7v2-validation = { version = "1.2.0", path = "../hl7v2-validation", optional = true }
-hl7v2-datatype = { version = "1.2.0", path = "../hl7v2-datatype", optional = true }
 hl7v2-datetime = { version = "1.2.0", path = "../hl7v2-datetime", optional = true }
 hl7v2-batch = { version = "1.2.0", path = "../hl7v2-batch", optional = true }
 hl7v2-stream = { version = "1.2.0", path = "../hl7v2-stream", optional = true }
